adaptive_rag/requirements.txt
langchain
langchain-openai
langchain-community
langgraph
faiss-cpu
google-search-results
python-dotenv
tiktoken
